refactor: add code that works on windows.
