+++ .gitignore
@@ -1,8 +1,2 @@
-.vscode
-x64
 .vs
-chip8.sln
-chip8.vcxproj
-chip8.vcxproj.filters
-chip8.vcxproj.user
-build
+x64
