+++ chip8-win/chip8-win.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.6.33723.286
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "chip8-win", "chip8-win.vcxproj", "{906673EB-452C-4EBE-9D21-ADDC8D1C8BF8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{906673EB-452C-4EBE-9D21-ADDC8D1C8BF8}.Debug|x64.ActiveCfg = Debug|x64
+		{906673EB-452C-4EBE-9D21-ADDC8D1C8BF8}.Debug|x64.Build.0 = Debug|x64
+		{906673EB-452C-4EBE-9D21-ADDC8D1C8BF8}.Debug|x86.ActiveCfg = Debug|Win32
+		{906673EB-452C-4EBE-9D21-ADDC8D1C8BF8}.Debug|x86.Build.0 = Debug|Win32
+		{906673EB-452C-4EBE-9D21-ADDC8D1C8BF8}.Release|x64.ActiveCfg = Release|x64
+		{906673EB-452C-4EBE-9D21-ADDC8D1C8BF8}.Release|x64.Build.0 = Release|x64
+		{906673EB-452C-4EBE-9D21-ADDC8D1C8BF8}.Release|x86.ActiveCfg = Release|Win32
+		{906673EB-452C-4EBE-9D21-ADDC8D1C8BF8}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {18A57BA4-764B-4DFD-AE4E-9353CAFE13B5}
+	EndGlobalSection
+EndGlobal
